refactor: 코드를 깔끔하게 수정
- styles/colors 파일만들어서 색 관리
- services 폴더 만들어서 각 로직들 관리
- useSnakeGameHook에서 있던 키보드 이벤트 핸들러를 SnakeGameCanvas로 이동
