+++ src/constants/snake.ts
@@ -13,9 +13,6 @@ export default Object.freeze({
   FRAME_START: 65,
   TICK_TOC_UNIT: 10,
 
-  SNAKE_BODY_COLOR: '#f5046d',
-  FOOD_COLOR: '#676FA3',
-
   SNAKE_LEFT_DIRECTION: 'LEFT',
   SNAKE_RIGHT_DIRECTION: 'RIGHT',
   SNAKE_UP_DIRECTION: 'UP',
