+++ src/styles/colors.ts
@@ -1,0 +1,3 @@
+export const SNAKE_COLOR = Object.freeze({
+  BODY_COLOR: '#f5046d',
+});
