+++ src/types/common.ts
@@ -1,0 +1,2 @@
+// TODO: Space 지우기
+export type KeyboardCodeType = 'ArrowRight' | 'ArrowLeft' | 'ArrowUp' | 'ArrowDown' | 'Space';
